requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
pytz~=2020.1
python-dateutil~=2.8.1
pync~=2.0.3